hotfix : 개발 진행 중 코드로 인한 오류 발생부분 주석 처리
Former-commit-id: ed3bc0de31e51b6060d4c8dd124a7bc59771f61d

+++ data/Table_create_query.sql
@@ -36,7 +36,7 @@ CREATE TABLE `Animal` (
 	`color`	varchar(32)	not NULL,
 	`sexcd`  CHAR(1) not NULL DEFAULT 'Q' COMMENT 'F : 암컷 M : 수컷 Q : 미상',
 	`age`	INT NOT NULL DEFAULT 1,
-	`weight`	INT DEFAULT 1,
+	`weight`	float DEFAULT 1,
 	`specialMark`	varchar(255)	NULL,
 	`neutral`	CHAR(1)	NULL COMMENT 'Y: 예 N : 아니오 U : 미상 ',
 	`noticeNo`	varchar(255)	NULL,
